config: refactor plugin related parts
